requirements.txt
# Data
pandas>=2.0.0
kaggle>=1.6.0

# NLP & ML - Baseline
scikit-learn>=1.3.0
nltk>=3.8.0

# NLP & ML - Deep Learning
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0  # LoRA
datasets>=2.16.0

# Experiment Tracking
mlflow>=2.9.0

# Evaluation
seaborn>=0.13.0
matplotlib>=3.8.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
PyYAML>=6.0.0
